db-bench: use feature flag for plotting
